refactor(D-170 D): decompose local_graph into sanitizers + explanation modules
1,769→1,193 lines. Two cohesive verbatim extractions with backward re-export:
- local_graph_sanitizers.py: foreign-script/meta-narration cleaning + dedup/
  near-dup + precision guardrail (pure string ops, zero module deps)
- local_graph_explanation.py: exercise-explanation-with-context + question-aware
  latency budgets (D-053); _format_history via late-binding (D-168) to avoid the
  circular import with local_graph's re-export

All internal callers + external importers (local_fallback, bac_exercise_skill)
untouched via re-export. Gate-pinned code stays in local_graph.py: the doctrine
import + _EXERCISE_EXPLANATION_SYSTEM_PROMPT (D-071 reads the file directly),
_apply_answer_quality_skill (D-073), the mirrored D-013 greeting patterns.

Companion cleanup: ruff auto-fix removed a redundant inline 'import contextlib'
in _supervisor_node, surfacing a latent F823 (module-level contextlib shadowed by
a later inline import in the same function) — removed the stray inline import so
the module-level one at line 27 resolves globally.

1576 services tests pass (+1 = realigned ISS-110). All fitness gates green;
runtime_truth lock regenerated for the two new modules.
